+++ projects/SelfTest/GeneratorTests.cpp
@@ -12,7 +12,7 @@
 
 #include "catch.hpp"
 
-inline size_t multiply( size_t a, size_t b )
+inline int multiply( int a, int b )
 {
     return a*b;
 }
@@ -21,8 +21,8 @@ TEST_CASE( "./succeeding/generators/1", "Generators over two ranges" )
 {
     using namespace Catch::Generators;
     
-    size_t i = GENERATE( between( 1, 5 ).then( values( 15, 20, 21 ).then( 36 ) ) );
-    size_t j = GENERATE( between( 100, 107 ) );
+    int i = GENERATE( between( 1, 5 ).then( values( 15, 20, 21 ).then( 36 ) ) );
+    int j = GENERATE( between( 100, 107 ) );
     
     REQUIRE( multiply( i, 2 ) == i*2 );
     REQUIRE( multiply( j, 2 ) == j*2 );
